refactor: prepare regexp extension
