+++ Makefile
@@ -34,7 +34,7 @@ compile-linux:
 	gcc -O3 $(LINIX_FLAGS) src/sqlite3-ipaddr.c src/ipaddr/*.c -o dist/ipaddr.so
 	gcc -O3 $(LINIX_FLAGS) src/sqlite3-json1.c -o dist/json1.so
 	gcc -O3 $(LINIX_FLAGS) src/sqlite3-math.c -o dist/math.so -lm
-	gcc -O3 $(LINIX_FLAGS) -DPCRE2_CODE_UNIT_WIDTH=8 -DLINK_SIZE=2 -DHAVE_CONFIG_H -DSUPPORT_UNICODE src/sqlite3-regexp.c src/regexp/regexp.c src/regexp/pcre2/*.c -o dist/regexp.so
+	gcc -O3 $(LINIX_FLAGS) -DPCRE2_CODE_UNIT_WIDTH=8 -DLINK_SIZE=2 -DHAVE_CONFIG_H -DSUPPORT_UNICODE src/sqlite3-regexp.c src/regexp/*.c src/regexp/pcre2/*.c -o dist/regexp.so
 	gcc -O3 $(LINIX_FLAGS) src/sqlite3-stats.c -o dist/stats.so -lm
 	gcc -O3 $(LINIX_FLAGS) src/sqlite3-text.c src/text/*.c -o dist/text.so
 	gcc -O3 $(LINIX_FLAGS) src/sqlite3-unicode.c -o dist/unicode.so
@@ -51,7 +51,7 @@ compile-windows:
 	gcc -O1 $(WINDO_FLAGS) -I. src/sqlite3-fuzzy.c src/fuzzy/*.c -o dist/fuzzy.dll
 	gcc -O3 $(WINDO_FLAGS) -I. src/sqlite3-json1.c -o dist/json1.dll
 	gcc -O3 $(WINDO_FLAGS) -I. src/sqlite3-math.c -o dist/math.dll -lm
-	gcc -O3 $(WINDO_FLAGS) -DPCRE2_CODE_UNIT_WIDTH=8 -DLINK_SIZE=2 -DHAVE_CONFIG_H -DSUPPORT_UNICODE -DPCRE2_STATIC -I. src/sqlite3-regexp.c src/regexp/regexp.c src/regexp/pcre2/*.c -o dist/regexp.dll
+	gcc -O3 $(WINDO_FLAGS) -DPCRE2_CODE_UNIT_WIDTH=8 -DLINK_SIZE=2 -DHAVE_CONFIG_H -DSUPPORT_UNICODE -DPCRE2_STATIC -I. src/sqlite3-regexp.c src/regexp/*.c src/regexp/pcre2/*.c -o dist/regexp.dll
 	gcc -O3 $(WINDO_FLAGS) -I. src/sqlite3-stats.c -o dist/stats.dll -lm
 	gcc -O3 $(WINDO_FLAGS) -I. src/sqlite3-text.c src/text/*.c -o dist/text.dll
 	gcc -O3 $(WINDO_FLAGS) -I. src/sqlite3-unicode.c -o dist/unicode.dll
@@ -69,7 +69,7 @@ compile-macos:
 	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-ipaddr.c src/ipaddr/*.c -o dist/ipaddr.dylib
 	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-json1.c -o dist/json1.dylib
 	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-math.c -o dist/math.dylib -lm
-	gcc -O3 $(MACOS_FLAGS) -DPCRE2_CODE_UNIT_WIDTH=8 -DLINK_SIZE=2 -DHAVE_CONFIG_H -DSUPPORT_UNICODE -I src src/sqlite3-regexp.c src/regexp/regexp.c src/regexp/pcre2/*.c -o dist/regexp.dylib
+	gcc -O3 $(MACOS_FLAGS) -DPCRE2_CODE_UNIT_WIDTH=8 -DLINK_SIZE=2 -DHAVE_CONFIG_H -DSUPPORT_UNICODE -I src src/sqlite3-regexp.c src/regexp/*.c src/regexp/pcre2/*.c -o dist/regexp.dylib
 	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-stats.c -o dist/stats.dylib -lm
 	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-text.c src/text/*.c -o dist/text.dylib
 	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-unicode.c -o dist/unicode.dylib
@@ -85,7 +85,7 @@ compile-macos-x86:
 	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-ipaddr.c src/ipaddr/*.c -o dist/x86/ipaddr.dylib -target x86_64-apple-macos10.12
 	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-json1.c -o dist/x86/json1.dylib -target x86_64-apple-macos10.12
 	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-math.c -o dist/x86/math.dylib -target x86_64-apple-macos10.12 -lm
-	gcc -O3 $(MACOS_FLAGS) -DPCRE2_CODE_UNIT_WIDTH=8 -DLINK_SIZE=2 -DHAVE_CONFIG_H -DSUPPORT_UNICODE -I src src/sqlite3-regexp.c src/regexp/regexp.c src/regexp/pcre2/*.c -o dist/x86/regexp.dylib -target x86_64-apple-macos10.12
+	gcc -O3 $(MACOS_FLAGS) -DPCRE2_CODE_UNIT_WIDTH=8 -DLINK_SIZE=2 -DHAVE_CONFIG_H -DSUPPORT_UNICODE -I src src/sqlite3-regexp.c src/regexp/*.c src/regexp/pcre2/*.c -o dist/x86/regexp.dylib -target x86_64-apple-macos10.12
 	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-stats.c -o dist/x86/stats.dylib -target x86_64-apple-macos10.12 -lm
 	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-text.c src/text/*.c -o dist/x86/text.dylib -target x86_64-apple-macos10.12
 	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-unicode.c -o dist/x86/unicode.dylib -target x86_64-apple-macos10.12
@@ -101,7 +101,7 @@ compile-macos-arm64:
 	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-ipaddr.c src/ipaddr/*.c -o dist/arm64/ipaddr.dylib -target arm64-apple-macos11
 	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-json1.c -o dist/arm64/json1.dylib -target arm64-apple-macos11
 	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-math.c -o dist/arm64/math.dylib -target arm64-apple-macos11 -lm
-	gcc -O3 $(MACOS_FLAGS) -DPCRE2_CODE_UNIT_WIDTH=8 -DLINK_SIZE=2 -DHAVE_CONFIG_H -DSUPPORT_UNICODE -I src src/sqlite3-regexp.c src/regexp/regexp.c src/regexp/pcre2/*.c -o dist/arm64/regexp.dylib -target arm64-apple-macos11
+	gcc -O3 $(MACOS_FLAGS) -DPCRE2_CODE_UNIT_WIDTH=8 -DLINK_SIZE=2 -DHAVE_CONFIG_H -DSUPPORT_UNICODE -I src src/sqlite3-regexp.c src/regexp/*.c src/regexp/pcre2/*.c -o dist/arm64/regexp.dylib -target arm64-apple-macos11
 	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-stats.c -o dist/arm64/stats.dylib -target arm64-apple-macos11 -lm
 	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-text.c src/text/*.c -o dist/arm64/text.dylib -target arm64-apple-macos11
 	gcc -O3 $(MACOS_FLAGS) -I src src/sqlite3-unicode.c -o dist/arm64/unicode.dylib -target arm64-apple-macos11
